init

Initiate the package by loading the data into parent frame. This should be done as soon as the package is loaded or attached. This creates(rewrites) new variables 'deptable' and 'packmeta' to the environment where it is run from.

init(local = FALSE, repository = "CRAN", ...)

Arguments

local
(flag, default: FALSE) If

FALSE: Tries to to download package data from CRAN over internet and compute dependencies

TRUE: Loads data that comes with the package corresponding to 2nd September 2017 02:04 IST

repository
(character vector, Default: "CRAN") One among c("CRAN", "BioCsoft", "BioCann", "BioCexp", "BioCextra", "omegahat"). To use a repository not in this list, set 'repository' to NULL and pass named argument called 'repos' with a valid repository address. This will be passed as is to `utils::available.packages()`.

...
Additional parameters to be passed to `available.packages()`

Value

An invisible TRUE